move diag under p2p/diag
