+++ p2p/diagnostics/internal/pb/diagnostics.proto
