+++ p2p/diagnostics/vis.go
